refactor : 프로젝트 폴더 구조 네이밍 및 구조 변경
