+++ src/main/java/com/ticket/concert/application/AuthService.java
@@ -1,12 +1,12 @@
-package com.ticket.concert.service;
+package com.ticket.concert.application;
 
 import com.ticket.concert.domain.LoginUser;
-import com.ticket.concert.domain.User;
-import com.ticket.concert.dto.auth.request.LoginRequest;
-import com.ticket.concert.dto.auth.response.LoginResponse;
+import com.ticket.concert.domain.user.User;
+import com.ticket.concert.presentation.dto.auth.request.LoginRequest;
+import com.ticket.concert.presentation.dto.auth.response.LoginResponse;
 import com.ticket.concert.global.exception.BusinessException;
 import com.ticket.concert.global.exception.constant.ErrorCode;
-import com.ticket.concert.repository.user.UserRepository;
+import com.ticket.concert.domain.user.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
