+++ src/main/java/com/ticket/concert/application/UserService.java
@@ -1,10 +1,10 @@
-package com.ticket.concert.service;
+package com.ticket.concert.application;
 
-import com.ticket.concert.domain.User;
-import com.ticket.concert.dto.user.request.JoinRequest;
+import com.ticket.concert.domain.user.User;
+import com.ticket.concert.presentation.dto.user.request.JoinRequest;
 import com.ticket.concert.global.exception.BusinessException;
 import com.ticket.concert.global.exception.constant.ErrorCode;
-import com.ticket.concert.repository.user.UserRepository;
+import com.ticket.concert.domain.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
