+++ src/main/java/com/ticket/concert/presentation/UserController.java
@@ -1,8 +1,8 @@
-package com.ticket.concert.controller;
+package com.ticket.concert.presentation;
 
-import com.ticket.concert.dto.user.request.JoinRequest;
+import com.ticket.concert.presentation.dto.user.request.JoinRequest;
 import com.ticket.concert.global.common.ApiResponse;
-import com.ticket.concert.service.UserService;
+import com.ticket.concert.application.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
