+++ src/main/java/com/ticket/concert/presentation/dto/user/request/JoinRequest.java
@@ -1,6 +1,6 @@
-package com.ticket.concert.dto.user.request;
+package com.ticket.concert.presentation.dto.user.request;
 
-import com.ticket.concert.domain.User;
+import com.ticket.concert.domain.user.User;
 import com.ticket.concert.global.regex.UserRegex;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
