+++ src/test/java/com/ticket/concert/domain/UserTest.java
@@ -1,6 +1,7 @@
 package com.ticket.concert.domain;
 
 import com.ticket.concert.domain.constant.Role;
+import com.ticket.concert.domain.user.User;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
